@@ -1,34 +1,33 @@
 ﻿using System;
 
-namespace AutofacExperiment.Access.Deamon
+namespace AutofacExperiment.Access.Deamon;
+
+/// <summary>
+///     Deamon resource repository.
+/// </summary>
+/// <seealso cref="AutofacExperiment.Access.Deamon.IDeamonResourceRepository" />
+/// <seealso cref="System.IDisposable" />
+public class DeamonResourceRepository : IDeamonResourceRepository
 {
     /// <summary>
-    /// Deamon resource repository.
+    ///     Gets the name.
     /// </summary>
-    /// <seealso cref="NineYi.ERP.DA.ERPDB.Deamon.IDeamonResourceRepository"/>
-    /// <seealso cref="System.IDisposable"/>
-    public class DeamonResourceRepository : IDeamonResourceRepository
+    /// <value>The name.</value>
+    public string Name => typeof(DeamonResourceRepository).FullName;
+
+    /// <summary>
+    ///     執行與釋放 (Free)、釋放 (Release) 或重設 Unmanaged 資源相關聯之應用程式定義的工作。
+    /// </summary>
+    public void Dispose()
     {
-        /// <summary>
-        /// Gets the name.
-        /// </summary>
-        /// <value>The name.</value>
-        public string Name { get { return typeof(DeamonResourceRepository).FullName; } }
+        Console.WriteLine("{0} dispose!", Name);
+    }
 
-        /// <summary>
-        /// 執行與釋放 (Free)、釋放 (Release) 或重設 Unmanaged 資源相關聯之應用程式定義的工作。
-        /// </summary>
-        public void Dispose()
-        {
-            Console.WriteLine(string.Format("{0} dispose!", this.Name));
-        }
-
-        /// <summary>
-        /// Eat resource
-        /// </summary>
-        public void ResourceMonster()
-        {
-            Console.WriteLine(string.Format("Object HashCode:{0}", this.GetHashCode()));
-        }
+    /// <summary>
+    ///     Eat resource
+    /// </summary>
+    public void ResourceMonster()
+    {
+        Console.WriteLine("Object HashCode:{0}", GetHashCode());
     }
 }
