@@ -1,15 +1,14 @@
 ﻿using System;
 
-namespace AutofacExperiment.Access.Deamon
+namespace AutofacExperiment.Access.Deamon;
+
+/// <summary>
+///     Iinterface of deamon resource repository
+/// </summary>
+public interface IDeamonResourceRepository : IDisposable
 {
     /// <summary>
-    /// Iinterface of deamon resource repository
+    ///     Eat resource
     /// </summary>
-    public interface IDeamonResourceRepository : IDisposable
-    {
-        /// <summary>
-        /// Eat resource
-        /// </summary>
-        void ResourceMonster();
-    }
+    void ResourceMonster();
 }
