@@ -1,47 +1,47 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Reflection;
+﻿using System.Reflection;
 using Autofac;
 using AutofacExperiment.Access.Deamon;
+using Module = Autofac.Module;
 
-namespace AutofacExperiment.Access.Modules
+namespace AutofacExperiment.Access.Modules;
+
+public class DAModules : Module
 {
-    public class DAModules : Autofac.Module
+    /// <summary>
+    ///     Override to add registrations to the container.
+    /// </summary>
+    /// <param name="builder">
+    ///     The builder through which components can be
+    ///     registered.
+    /// </param>
+    /// <remarks>
+    ///     Note that the ContainerBuilder parameter is unique to this module.
+    ///     ref：http://autofac.readthedocs.io/en/latest/lifetime/instance-scope.html
+    /// </remarks>
+    protected override void Load(ContainerBuilder builder)
     {
-        /// <summary>
-        /// Override to add registrations to the container.
-        /// </summary>
-        /// <param name="builder">The builder through which components can be
-        /// registered.</param>
-        /// <remarks>
-        /// Note that the ContainerBuilder parameter is unique to this module.
-        /// 參考資料：http://autofac.readthedocs.io/en/latest/lifetime/instance-scope.html
-        /// </remarks>
-        protected override void Load(ContainerBuilder builder)
-        {
-            var assembly = Assembly.Load("NineYi.ERP.DA.ERPDB");
+        var assembly = Assembly.Load("AutofacExperiment.Access");
 
-            //builder.RegisterAssemblyTypes(assembly).OwnedByLifetimeScope();
+        builder.RegisterAssemblyTypes(assembly).OwnedByLifetimeScope();
 
-            //// 目前NMQV2的物件生命週期policy
-            builder.RegisterType<DeamonResourceRepository>()
-                .As<IDeamonResourceRepository>()
-                .InstancePerDependency();
+        //// 目前 Deamon 的物件生命週期 policy
+        builder.RegisterType<DeamonResourceRepository>()
+            .As<IDeamonResourceRepository>()
+            .InstancePerDependency();
 
-            //builder.RegisterType<DeamonResourceRepository>()
-            //    .As<IDeamonResourceRepository>()
-            //    .OwnedByLifetimeScope();
+        //builder.RegisterType<DeamonResourceRepository>()
+        //    .As<IDeamonResourceRepository>()
+        //    .OwnedByLifetimeScope();
 
-            //// 這個生命週期讓 MemoryLeakMethod02 不會有 memory leak 的問題
-            //// 因為在同一個scope下取得的實體，將會是同一個實體。(新的scope就是新的實體)
-            //builder.RegisterType<DeamonResourceRepository>()
-            //    .As<IDeamonResourceRepository>()
-            //    .InstancePerLifetimeScope();
+        //// 這個生命週期讓 MemoryLeakMethod02 不會有 memory leak 的問題
+        //// 因為在同一個scope下取得的實體，將會是同一個實體。(新的scope就是新的實體)
+        //builder.RegisterType<DeamonResourceRepository>()
+        //    .As<IDeamonResourceRepository>()
+        //    .InstancePerLifetimeScope();
 
-            ////用這個policy就不用擔心 memory leak 的問題
-            //builder.RegisterType<DeamonResourceRepository>()
-            //    .As<IDeamonResourceRepository>()
-            //    .SingleInstance();    
-        }
+        ////用這個policy就不用擔心 memory leak 的問題
+        //builder.RegisterType<DeamonResourceRepository>()
+        //    .As<IDeamonResourceRepository>()
+        //    .SingleInstance();    
     }
 }
